an alternative approach
This way is much more theoretically elegant on account of treating the
operations of placing-words and "placing strings that form prefixes of
words in the orthogonal direction" as equals. From a few cursory test
runs, it's been observed to solve some cases extremely fast compared to
the earlier algorithm (seeded with TAIL in the first row, it found
[[:T :A :I :L] [:I :R :M :A] [:D :I :P :S] [:E :A :S :T]] in just 16
seconds!), it's actually not clearly better on average. Further research
is needed.