simple benchmarking tools
It's looking like the old `solve` algorithm is better than the new
`solve-it` on average: 11.6 and 21.5 and 12.7 vs. 17.6 seconds
respectively on two samples of ten 3-by-3 grids seeded with a
random (arbitrary human-chosen subvector of our hash-ordered wordlist)
word in the top row. If I saw some very impressive results from
`solve-it` the other night, maybe that means it's higher-variance?