confused attempt to prevent duplicate words in same grid
Probably because of hash-ordering and topological properties of word
sqaures that I don't understand well enough to articulate, our current
code is prone to reuse the same word in a grid in both the Across and
Down orientations at a rate far above "chance" (whatever that means),
which is less interesting than always using unique words. Here's an
extreme example:

[H A N N A]
[A L E U T]
[N E L D A]
[N U D E R]
[A T A R I]

I delusionally thought a small change to `admissibles` could remedy
this, but actually it doesn't: in the example above, placing ATARI in
the last row forms five (!) duplicates, but ATARI is still itself
admissible because it hasn't been previously placed. I'm leaving the
modification in for now because it is (should be) harmless and at least
prevents (should prevent) the same word being used twice in the Across
orientation (not that I've seen that come up).

Also: an alias for the keywordization of Unicodepoint 0x2588; unused,
and it may remain unused, and yet perhaps also a foreshadowing of code
yet to come?