`containing-address-down`, Worst Function Ever?
`wordspan-addresses-across` was already pretty bad, but this is just
abominable (but, passes tests). My hope is that further experience in
this endeavor will grant me insight into what abstractions I should be
using, which will enable me to rewrite some of the more unsightly
aspects of what we have so far.