running out of ideas ...
It's going to take another algorithmic insight (and maybe some profiling
tools more sophisticated than the occasional `println` or `time`), or a
lot more patience, to get this to generate puzzles on a "realistically"
sized grid in a "reasonable" (patience-limited) amount of time. The idea
about ordering the letter-tree children by their number of successors
sounded plausible but didn't seem to help (and maybe hurt a bit; I don't
know what JVM-magic happens under the hood, but it's hard to distinguish
a signal when my timings vary so much).